More Start Card Filtering
Cards that start with `ん` are filtered out of the starting cards because choosing one ends in a instant game over anyways.

Also I ended up getting 100% on my word catalog while working on this, lol. So I guess I can confirm it works well.

I struggled the most with the priconneyomi phrases hidden behind `たて`... If you have a chance to select one of these words, don't miss it.
`けん` is also a pretty bad one too, so I heard. it's only available if Kaya selects `おばけ`.